+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Fixed
+- Fix Jenkins log warning `null not assignable to interface hudson.model.Action`
 
 ## v1.0.2 - 2020-06-17
 ### Fixed