+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Update plugin test dependencies
 - Fix IllegalStateExceptions caused by multiple onSuccess/onComplete calls on newer workflow plugin
   versions
+- Fix missing Jenkins build logs on newer workflow plugin versions when run was stopped
 
 ## v1.0.2 - 2020-06-17
 ### Fixed